Update audio-files.md
I added three words, thinking that it might be a generally good thing (and I'm willing to be wrong about this) for a podcast listening app to automatically apply genuine "improvements" to a file before playing it back, like setting the LUFS to -16 (stereo, whatever for mono) and applying intelligent dynamic processing.

Again, I'm willing to be wrong on this. Furthermore, I think these settings should be ideally be set at the podcast hosting level. But in reality, hosting companies are slow to move.

Maybe a notice that the audio has been "enhanced"? And maybe storing the original version for playing upon request is another option?